requirements.txt
flask>=2.0.0
requests>=2.26.0
open3d>=0.16.0
natsort>=8.0.0
numpy>=1.21.0
scipy>=1.7.0
diffusers>=0.11.0
torch>=1.10.0
tqdm>=4.62.0
wandb>=0.12.0
huggingface-hub>=0.10.0
spconv-cu117>=2.3.0  # Adjust CUDA version as needed
opencv-python>=4.5.0
clean-fid>=0.1.0 